# server/src/middleware/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: 'POCAZ REST API 관리 페이지'
  description: |-
    POCAZ 내 client가 server에 요청하는 방법을 기술한 페이지입니다.

    서버 관리자는 본인이 담당한 API를 개발 후 현재 페이지에 기술해주시길 바랍니다. 감사합니다.
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:8000/
  - url: http://localhost:8080/
...
paths:
...
  /artist/poca:
    post:
      tags:
        - photocard
      summary: 신규 포토카드 입력
      description: 'DB에 없는 포토카드 등록'
      operationId: writePhotocard
      requestBody:
        description: 수정할 내용을 입력해요
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Photocard'
      responses:
        '200':
          description: 'successful operation'
    get:
      tags:
        - photocard
      summary: DB상 포토카드 전체 조회
      description: ''
      operationId: getPhotocards
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Photocard'
  /artist/poca/{id}:
    get:
      tags:
        - photocard
      summary: '포토카드 개별 조회'
      description: ''
      operationId: getPhotocard
      parameters:
        - name: id
          in: path
          description: category of post
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Photocard'
    put:
      tags:
        - photocard
      summary: 포토카드 정보 수정
      description: ''
      operationId: modifyPhotocard
      parameters:
        - name: id
          in: path
          description: id of photocard
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        description: 수정할 내용을 입력해요
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Photocard'
      responses:
        '200':
          description: 'successful operation'
    delete:
      tags:
        - photocard
      summary: 포토카드 제거( DB 수정 전까지는 사용하지 말것)
      description: ''
      operationId: deletePhotocard
      parameters:
        - name: id
          in: path
          description: 포토카드의 key 값
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: successfully erased!
        '400':
          description: Invalid post
        '404':
          description: Not Found
  /artist/galmang/{user}:
    post:
      tags:
        - galmangPoca
      summary: 유저 갈망포카 입력
      description: ''
      operationId: writeGalmang
      parameters:
        - name: user
          in: path
          description: 포토카드의 key 값
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        description: 내용을 입력해요
        content:
          application/json:
            schema:
              type: integer
              example: [1, 2]
      responses:
        '200':
          description: 'successful operation'
    get:
      tags:
        - galmangPoca
      summary: 유저 본인의 갈망포카 전체 조회
      description: ''
      operationId: 'getGalmang'
      parameters:
        - name: user
          in: path
          description: 포토카드의 key 값
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Galmang'
  /artist/galmang/{user}/{id}:
    delete:
      tags:
        - galmangPoca
      summary: 유저 갈망포카 제거
      description: ''
      operationId: deleteGalmang
      parameters:
        - name: user
          in: path
          description: 갈망포카 key 값
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: id
          in: path
          description: 갈망포카 key 값
          required: true
          schema:
            type: integer
            format: int64
          example: 2
      responses:
        '200':
          description: successfully erased!
        '400':
          description: Invalid post
        '404':
          description: Not Found
...
components:
  schemas:
...
    Photocard:
      type: object
      properties:
        artist:
          type: integer
          example: 1
        path:
          type: string
          example: 'https://www.naver.com'
        name:
          type: string
          example: '포카칩'
        description:
          type: string
          example: '좋은 포카입니다.'
    Galmang:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user:
          type: integer
          example: 1
        photocard:
          type: integer
          example: 1